requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.1.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asgiref=3.5.2=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.4.26=hecd8cb5_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py310hecd8cb5_0
cffi=1.15.0=pypi_0
chardet=3.0.4=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.1.3=pypi_0
colorama=0.4.4=pypi_0
cython=0.29.30=pypi_0
dataclasses=0.6=pypi_0
django=4.0.5=pypi_0
docopt=0.6.2=pypi_0
editdistance=0.6.0=pypi_0
fairseq=0.10.0=pypi_0
filelock=3.7.1=pypi_0
flatbuffers=1.12=pypi_0
fugashi=1.1.2=pypi_0
gast=0.4.0=pypi_0
google-auth=2.8.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googletrans=3.0.0=pypi_0
grpcio=1.46.3=pypi_0
h11=0.9.0=pypi_0
h2=3.2.0=pypi_0
h5py=3.7.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2021.12.1=pypi_0
httpcore=0.9.1=pypi_0
httpx=0.13.3=pypi_0
huggingface-hub=0.7.0=pypi_0
hydra-core=1.2.0=pypi_0
hyperframe=5.2.0=pypi_0
idna=2.10=pypi_0
ipadic=1.0.0=pypi_0
joblib=1.1.0=pypi_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
libclang=14.0.1=pypi_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
markdown=3.3.7=pypi_0
mecab-python3=1.0.5=pypi_0
mosestokenizer=1.2.1=pypi_0
ncurses=6.3=hca72f7f_2
nltk=3.7=pypi_0
numpy=1.22.4=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.2.2=pypi_0
openfile=0.0.7=pypi_0
openssl=1.1.1q=hca72f7f_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pip=21.2.4=py310hecd8cb5_0
portalocker=2.4.0=pypi_0
portaudio=19.6.0=h647c56a_4
protobuf=3.19.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyaudio=0.2.11=pypi_0
pycparser=2.21=pypi_0
pyparsing=3.0.9=pypi_0
python=3.10.4=hdfd78df_0
pyyaml=6.0=pypi_0
readline=8.1.2=hca72f7f_1
regex=2022.6.2=pypi_0
requests=2.28.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.8=pypi_0
sacrebleu=2.1.0=pypi_0
sacremoses=0.0.53=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=61.2.0=py310hecd8cb5_0
six=1.16.0=pypi_0
sniffio=1.2.0=pypi_0
speechrecognition=3.8.1=pypi_0
sqlite=3.38.3=h707629a_0
sqlparse=0.4.2=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.9.1=pypi_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.12.1=pypi_0
toolwrapper=2.1.0=pypi_0
torch=1.11.0=pypi_0
tqdm=4.64.0=pypi_0
transformers=4.19.4=pypi_0
typing-extensions=4.2.0=pypi_0
tzdata=2022a=hda174b7_0
uctools=1.3.0=pypi_0
unidic-lite=1.0.8=pypi_0
urllib3=1.26.9=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.5=hca72f7f_1
zlib=1.2.12=h4dc903c_2
